+++ packages/cli/src/commands/index.ts
@@ -19,3 +19,4 @@ export { default as commandSiteKeyRegister } from "./siteKeyRegister.js";
 export { default as commandSiteKeyRegisterApi } from "./siteKeyRegisterApi.js";
 export { default as commandEnsureExternalIndexes } from "./ensureExternalIndexes.js";
 export { default as commandEnsureIndexes } from "./ensureIndexes.js";
+export { default as commandMigrateAbuserScoreThreshold } from "./migrateAbuserScoreThreshold.js";