+++ packages/provider/src/api/verify.ts
@@ -127,6 +127,8 @@ export function prosopoVerifyRouter(env: ProviderEnvironment): Router {
 						userAccessRulesStorage,
 						email,
 						clientRecord.settings.spamEmailDomainCheckEnabled,
+						clientRecord.settings.spamFilter,
+						clientRecord.settings.trafficFilter,
 					);
 
 				req.logger.debug(() => ({ data: { response } }));
@@ -137,6 +139,7 @@ export function prosopoVerifyRouter(env: ProviderEnvironment): Router {
 						req.i18n.t,
 						response[ApiParams.score],
 						response[ApiParams.commitmentId],
+						response[ApiParams.status],
 					);
 				res.json(verificationResponse);
 			} catch (err) {
@@ -229,7 +232,7 @@ export function prosopoVerifyRouter(env: ProviderEnvironment): Router {
 				// Will throw an error if the signature is invalid
 				verifySignature(dappSignature, timestamp.toString(), dappPair);
 
-				const { verified, score } =
+				const { verified, score, reason } =
 					await tasks.powCaptchaManager.serverVerifyPowCaptchaSolution(
 						dapp,
 						challenge,
@@ -239,6 +242,8 @@ export function prosopoVerifyRouter(env: ProviderEnvironment): Router {
 						userAccessRulesStorage,
 						email,
 						clientRecord.settings.spamEmailDomainCheckEnabled,
+						clientRecord.settings.spamFilter,
+						clientRecord.settings.trafficFilter,
 					);
 
 				const verificationResponse: VerificationResponse =
@@ -247,6 +252,7 @@ export function prosopoVerifyRouter(env: ProviderEnvironment): Router {
 						clientRecord,
 						req.i18n.t,
 						score,
+						reason,
 					);
 
 				return res.json(verificationResponse);