+++ demos/cypress-shared/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @prosopo/cypress-shared
 
+## 2.8.69
+### Patch Changes
+
+- Updated dependencies [3e54c0a]
+  - @prosopo/types@3.15.0
+
+## 2.8.68
+### Patch Changes
+
+- Updated dependencies [946a8ba]
+- Updated dependencies [5614814]
+  - @prosopo/types@3.14.1
+
+## 2.8.67
+### Patch Changes
+
+- Updated dependencies [fc514dd]
+- Updated dependencies [42650db]
+  - @prosopo/types@3.14.0
+
 ## 2.8.66
 ### Patch Changes