+++ integration/frameworks/svelte/svelte-procaptcha-integration-demo/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @prosopo/svelte-procaptcha-integration-demo
 
+## 1.1.65
+### Patch Changes
+
+  - @prosopo/svelte-procaptcha-wrapper@1.1.65
+
+## 1.1.64
+### Patch Changes
+
+  - @prosopo/svelte-procaptcha-wrapper@1.1.64
+
+## 1.1.63
+### Patch Changes
+
+  - @prosopo/svelte-procaptcha-wrapper@1.1.63
+
 ## 1.1.62
 ### Patch Changes