+++ integration/frameworks/vue/vue-procaptcha-wrapper/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @prosopo/vue-procaptcha-wrapper
 
+## 1.1.65
+### Patch Changes
+
+  - @prosopo/procaptcha-wrapper@2.6.67
+
+## 1.1.64
+### Patch Changes
+
+  - @prosopo/procaptcha-wrapper@2.6.66
+
+## 1.1.63
+### Patch Changes
+
+  - @prosopo/procaptcha-wrapper@2.6.65
+
 ## 1.1.62
 ### Patch Changes