+++ packages/procaptcha-react/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @prosopo/procaptcha-react
 
+## 2.9.52
+### Patch Changes
+
+- Updated dependencies [3e54c0a]
+  - @prosopo/types@3.15.0
+  - @prosopo/procaptcha@2.10.13
+  - @prosopo/procaptcha-common@2.10.3
+
+## 2.9.51
+### Patch Changes
+
+- Updated dependencies [946a8ba]
+- Updated dependencies [5614814]
+- Updated dependencies [b94890c]
+  - @prosopo/types@3.14.1
+  - @prosopo/locale@3.2.1
+  - @prosopo/common@3.1.33
+  - @prosopo/procaptcha@2.10.12
+  - @prosopo/procaptcha-common@2.10.2
+
+## 2.9.50
+### Patch Changes
+
+- Updated dependencies [06970d2]
+- Updated dependencies [fc514dd]
+- Updated dependencies [42650db]
+- Updated dependencies [dd3e06e]
+  - @prosopo/procaptcha-common@2.10.1
+  - @prosopo/widget-skeleton@2.8.1
+  - @prosopo/locale@3.2.0
+  - @prosopo/types@3.14.0
+  - @prosopo/procaptcha@2.10.11
+  - @prosopo/common@3.1.32
+
+## 2.9.49
+### Patch Changes
+
+- Updated dependencies [73df23c]
+- Updated dependencies [8139819]
+- Updated dependencies [4a9c518]
+  - @prosopo/procaptcha-common@2.10.0
+  - @prosopo/widget-skeleton@2.8.0
+  - @prosopo/common@3.1.31
+  - @prosopo/procaptcha@2.10.10
+
 ## 2.9.48
 ### Patch Changes